requirements.txt
numpy~=2.2.0
scipy~=1.14.1
pandas~=2.2.3
mysql.connector